Forcing character encoding to UTF-8
On Linux 64-bit, it seems that the ruby interpreter is not picking up the `-E UTF-8` command line switch in the wrapper shell script.This was causing issues in our pact tests on Linux, where apparently identical strings were failing provider verification, when non-ASCII whitespace was appearing in the text. The same test works fine on OSX.

Changing the switch to the somewhat more obscure `-Ku` switch seems to set the encoding correctly. See `man ruby` for more information on this switch.